objstore: Added common acceptance test for common objstore behaviour. (#327)
Additionally:
- Fixed inmen storage.
- Added leaktest where we can.
- Added ForeachObjStore test suite.

Currently tested by CI: GCS (only for contributors with write perm) and inmem only
