+++ pkg/objstore/inmem/inmem.go
@@ -9,8 +9,11 @@ import (
 	"io/ioutil"
 	"strings"
 
+	"github.com/improbable-eng/thanos/pkg/objstore"
 	"github.com/pkg/errors"
 )
+
+var errNotFound = errors.New("inmem: object not found")
 
 // Bucket implements the store.Bucket and shipper.Bucket interfaces against local memory.
 type Bucket struct {
@@ -34,18 +37,40 @@ func (b *Bucket) Objects() map[string][]byte {
 func (b *Bucket) Iter(_ context.Context, dir string, f func(string) error) error {
 	unique := map[string]struct{}{}
 
-	for filename := range b.objects {
-		if !strings.HasPrefix(filename, dir) {
+	var dirPartsCount int
+	dirParts := strings.SplitAfter(dir, objstore.DirDelim)
+	for _, p := range dirParts {
+		if p == "" {
 			continue
 		}
-		parts := strings.SplitAfter(filename, "/")
-		unique[parts[0]] = struct{}{}
+		dirPartsCount++
 	}
+	for filename := range b.objects {
+		if !strings.HasPrefix(filename, dir) || dir == filename {
+			continue
+		}
+
+		parts := strings.SplitAfter(filename, objstore.DirDelim)
+		unique[strings.Join(parts[:dirPartsCount+1], "")] = struct{}{}
+	}
+
 	var keys []string
 	for n := range unique {
 		keys = append(keys, n)
 	}
-	sort.Strings(keys)
+	sort.Slice(keys, func(i, j int) bool {
+		if strings.HasSuffix(keys[i], objstore.DirDelim) && strings.HasSuffix(keys[j], objstore.DirDelim) {
+			return strings.Compare(keys[i], keys[j]) < 0
+		}
+		if strings.HasSuffix(keys[i], objstore.DirDelim) {
+			return false
+		}
+		if strings.HasSuffix(keys[j], objstore.DirDelim) {
+			return true
+		}
+
+		return strings.Compare(keys[i], keys[j]) < 0
+	})
 
 	for _, k := range keys {
 		if err := f(k); err != nil {
@@ -57,9 +82,13 @@ func (b *Bucket) Iter(_ context.Context, dir string, f func(string) error) error
 
 // Get returns a reader for the given object name.
 func (b *Bucket) Get(_ context.Context, name string) (io.ReadCloser, error) {
+	if name == "" {
+		return nil, errors.New("inmem: object name is empty")
+	}
+
 	file, ok := b.objects[name]
 	if !ok {
-		return nil, errors.Errorf("no such file %s", name)
+		return nil, errNotFound
 	}
 
 	return ioutil.NopCloser(bytes.NewReader(file)), nil
@@ -67,13 +96,17 @@ func (b *Bucket) Get(_ context.Context, name string) (io.ReadCloser, error) {
 
 // GetRange returns a new range reader for the given object name and range.
 func (b *Bucket) GetRange(_ context.Context, name string, off, length int64) (io.ReadCloser, error) {
+	if name == "" {
+		return nil, errors.New("inmem: object name is empty")
+	}
+
 	file, ok := b.objects[name]
 	if !ok {
-		return nil, errors.Errorf("no such file %s", name)
+		return nil, errNotFound
 	}
 
 	if int64(len(file)) < off {
-		return nil, errors.Errorf("offset larger than content length. Len %d. Offset: %v", len(file), off)
+		return nil, errors.Errorf("inmem: offset larger than content length. Len %d. Offset: %v", len(file), off)
 	}
 
 	if int64(len(file)) <= off+length {
@@ -104,4 +137,9 @@ func (b *Bucket) Upload(_ context.Context, name string, r io.Reader) error {
 func (b *Bucket) Delete(_ context.Context, name string) error {
 	delete(b.objects, name)
 	return nil
+}
+
+// IsObjNotFoundErr returns true if error means that object is not found. Relevant to Get operations.
+func (b *Bucket) IsObjNotFoundErr(err error) bool {
+	return err == errNotFound
 }
